# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/logistics/order/{orderId}": {
      "get": {
        "summary": "查询订单物流信息",
        "deprecated": false,
        "description": "查询订单物流信息",
        "tags": [],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLogisticsVO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "orderId": 0,
                    "orderNo": "",
                    "logisticsCompany": "",
                    "trackingNo": "",
                    "transportType": "",
                    "departureTime": "",
                    "arrivalTime": "",
                    "status": "",
                    "traces": [
                      {
                        "id": 0,
                        "nodeTime": "",
                        "nodeLocation": "",
                        "nodeDesc": ""
                      }
                    ],
                    "createTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/payments/order/{orderId}": {
      "get": {
        "summary": "查询订单支付记录",
        "deprecated": false,
        "description": "查询订单支付记录",
        "tags": [],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListPaymentRecordVO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "orderId": 0,
                      "orderNo": "",
                      "paymentNo": "",
                      "paymentStage": "",
                      "amount": 0,
                      "paymentMethod": "",
                      "status": "",
                      "paymentTime": "",
                      "voucherUrl": "",
                      "createTime": ""
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/orders/page": {
      "get": {
        "summary": "分页查询订单",
        "deprecated": false,
        "description": "分页查询订单",
        "tags": [],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPagePurchaseOrder"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderNo": "",
                        "contractId": 0,
                        "productInfo": {
                          "": {}
                        },
                        "actualQuantity": 0,
                        "actualAmount": 0,
                        "status": "",
                        "deliveryTime": "",
                        "inspectionResult": "",
                        "createTime": "",
                        "updateTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/orders/my": {
      "get": {
        "summary": "查询我的订单",
        "deprecated": false,
        "description": "查询我的订单",
        "tags": [],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPagePurchaseOrder"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderNo": "",
                        "contractId": 0,
                        "productInfo": {
                          "": {}
                        },
                        "actualQuantity": 0,
                        "actualAmount": 0,
                        "status": "",
                        "deliveryTime": "",
                        "inspectionResult": "",
                        "createTime": "",
                        "updateTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "summary": "订单详情",
        "deprecated": false,
        "description": "订单详情",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPurchaseOrder"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "orderNo": "",
                    "contractId": 0,
                    "productInfo": {
                      "": {}
                    },
                    "actualQuantity": 0,
                    "actualAmount": 0,
                    "status": "",
                    "deliveryTime": "",
                    "inspectionResult": "",
                    "createTime": "",
                    "updateTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/statistics/purchaser/{purchaserId}": {
      "get": {
        "summary": "获取采购方采购统计",
        "deprecated": false,
        "description": "获取采购方采购统计",
        "tags": [],
        "parameters": [
          {
            "name": "purchaserId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPurchaserStats"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "demandCount": 0,
                    "ordersCount": 0,
                    "totalPurchase": 0,
                    "averageOrderAmount": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/purchaser-info/{id}": {
      "get": {
        "summary": "获取采购方详情",
        "deprecated": false,
        "description": "获取采购方详情",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPurchaserInfo"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "userId": 0,
                    "companyName": "",
                    "companyType": "",
                    "businessLicense": "",
                    "purchaseScale": "",
                    "preferredOrigin": [
                      ""
                    ],
                    "auditStatus": "",
                    "auditRemark": "",
                    "createTime": "",
                    "updateTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改采购方信息",
        "deprecated": false,
        "description": "修改采购方信息",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaserInfoRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/stock-reservation/order/{orderId}": {
      "get": {
        "summary": "获取订单的库存预留记录",
        "deprecated": false,
        "description": "获取订单的库存预留记录",
        "tags": [],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultStockReservation"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "productId": 0,
                    "orderId": 0,
                    "reservedQuantity": 0,
                    "status": "",
                    "releaseReason": "",
                    "expiredTime": "",
                    "createTime": "",
                    "updateTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "description": "",
            "type": "null"
          }
        }
      },
...
      "PurchaserInfoRequest": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "description": ""
          },
          "companyType": {
            "type": "string",
            "description": ""
          },
          "businessLicense": {
            "type": "string",
            "description": ""
          },
          "purchaseScale": {
            "type": "string",
            "description": ""
          },
          "preferredOrigin": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": ""
          }
        },
        "required": [
          "companyName",
          "businessLicense"
        ]
      },
...
      "LogisticsTraceVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "nodeTime": {
            "type": "string",
            "description": ""
          },
          "nodeLocation": {
            "type": "string",
            "description": ""
          },
          "nodeDesc": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "LogisticsVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": "关联订单编号"
          },
          "logisticsCompany": {
            "type": "string",
            "description": ""
          },
          "trackingNo": {
            "type": "string",
            "description": ""
          },
          "transportType": {
            "type": "string",
            "description": ""
          },
          "departureTime": {
            "type": "string",
            "description": ""
          },
          "arrivalTime": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "",
            "enum": [
              "PENDING",
              "SHIPPED",
              "TRANSIT",
              "ARRIVED",
              "SIGNED"
            ]
          },
          "traces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogisticsTraceVO",
              "description": "物流轨迹视图对象"
            },
            "description": "物流轨迹列表"
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ResultLogisticsVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/LogisticsVO",
            "description": ""
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string",
            "description": "需要进行排序的字段"
          },
          "asc": {
            "type": "boolean",
            "description": "是否正序排列，默认 true",
            "default": true
          }
        }
      },
...
      "PaymentRecordVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": "关联订单编号"
          },
          "paymentNo": {
            "type": "string",
            "description": ""
          },
          "paymentStage": {
            "type": "string",
            "description": ""
          },
          "amount": {
            "type": "number",
            "description": ""
          },
          "paymentMethod": {
            "type": "string",
            "description": "",
            "enum": [
              "BANK_TRANSFER",
              "ALIPAY",
              "WECHAT"
            ]
          },
          "status": {
            "type": "string",
            "description": "",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED"
            ]
          },
          "paymentTime": {
            "type": "string",
            "description": ""
          },
          "voucherUrl": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ResultListPaymentRecordVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRecordVO",
              "description": "支付记录视图对象"
            },
            "description": ""
          }
        }
      },
...
      "MapObject": {
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/key1"
          }
        }
      },
...
      "PurchaseOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": ""
          },
          "contractId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "productInfo": {
            "$ref": "#/components/schemas/MapObject",
            "description": ""
          },
          "actualQuantity": {
            "type": "integer",
            "description": ""
          },
          "actualAmount": {
            "type": "number",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "",
            "enum": [
              "PENDING",
              "DELIVERED",
              "PAID",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "deliveryTime": {
            "type": "string",
            "description": ""
          },
          "inspectionResult": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          },
          "updateTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "PagePurchaseOrder": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseOrder",
              "description": "采购订单实体类"
            },
            "description": "查询数据列表",
            "default": "Collections.emptyList()"
          },
          "total": {
            "type": "integer",
            "description": "总数",
            "format": "int64",
            "default": 0
          },
          "size": {
            "type": "integer",
            "description": "每页显示条数，默认 10",
            "format": "int64",
            "default": 10
          },
          "current": {
            "type": "integer",
            "description": "当前页",
            "format": "int64",
            "default": 1
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": "排序字段信息",
            "default": "new ArrayList<>()"
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": "自动优化 COUNT SQL",
            "default": true
          },
          "searchCount": {
            "type": "boolean",
            "description": "是否进行 count 查询",
            "default": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": "{@link #optimizeJoinOfCountSql()}",
            "default": true
          },
          "maxLimit": {
            "type": "integer",
            "description": "单页分页条数限制",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": "countId"
          }
        }
      },
...
      "ResultPagePurchaseOrder": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PagePurchaseOrder",
            "description": ""
          }
        }
      },
...
      "PurchaserStats": {
        "type": "object",
        "properties": {
          "demandCount": {
            "type": "integer",
            "description": "发布的需求数",
            "format": "int64"
          },
          "ordersCount": {
            "type": "integer",
            "description": "订单数",
            "format": "int64"
          },
          "totalPurchase": {
            "type": "number",
            "description": "总采购额"
          },
          "averageOrderAmount": {
            "type": "number",
            "description": "平均订单金额"
          }
        }
      },
...
      "ResultPurchaserStats": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PurchaserStats",
            "description": ""
          }
        }
      },
...
      "key1": {
        "type": "object",
        "properties": {}
      },
...
      "ResultPurchaseOrder": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PurchaseOrder",
            "description": ""
          }
        }
      },
      "PurchaserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "companyName": {
            "type": "string",
            "description": ""
          },
          "companyType": {
            "type": "string",
            "description": ""
          },
          "businessLicense": {
            "type": "string",
            "description": ""
          },
          "purchaseScale": {
            "type": "string",
            "description": ""
          },
          "preferredOrigin": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": ""
          },
          "auditStatus": {
            "type": "string",
            "description": "",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "auditRemark": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          },
          "updateTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ResultPurchaserInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PurchaserInfo",
            "description": ""
          }
        }
      },
...
      "StockReservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "description": "产品ID",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "reservedQuantity": {
            "type": "integer",
            "description": "预留数量"
          },
          "status": {
            "type": "string",
            "description": "预留状态：reserved(已预留) released(已释放) expired(已过期)"
          },
          "releaseReason": {
            "type": "string",
            "description": "释放原因"
          },
          "expiredTime": {
            "type": "string",
            "description": "预留过期时间"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间"
          }
        }
      },
      "ResultStockReservation": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/StockReservation",
            "description": ""
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}